:white_check_mark: GWT BehaviorSpec 이용에서 @Test 이용한 테스트로 수정
